model_pipeline/requirements.txt
pyyaml
ultralytics
mlflow
pandas
matplotlib
pytest